+++ app/src/main/java/com/example/yumecafeteria/data/model/Orders.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity
 data class Orders(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0
+    val id: Long = 0L
 )
